# backend/openapi.json
{
  "swagger": "2.0",
  "info": {
    "description": "",
    "title": "standard public schema",
    "version": "14.5"
  },
  "host": "gihitmxztbocbbdbgsqx.supabase.co:443",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json",
    "application/vnd.pgrst.object+json;nulls=stripped",
    "application/vnd.pgrst.object+json",
    "text/csv"
  ],
  "produces": [
    "application/json",
    "application/vnd.pgrst.object+json;nulls=stripped",
    "application/vnd.pgrst.object+json",
    "text/csv"
  ],
  "paths": {
...
    "/nodes": {
      "get": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.nodes.id"
          },
          {
            "$ref": "#/parameters/rowFilter.nodes.node_id"
          },
          {
            "$ref": "#/parameters/rowFilter.nodes.gateway_id"
          },
          {
            "$ref": "#/parameters/rowFilter.nodes.crop_name"
          },
          {
            "$ref": "#/parameters/rowFilter.nodes.created_at"
          },
          {
            "$ref": "#/parameters/select"
          },
          {
            "$ref": "#/parameters/order"
          },
          {
            "$ref": "#/parameters/range"
          },
          {
            "$ref": "#/parameters/rangeUnit"
          },
          {
            "$ref": "#/parameters/offset"
          },
          {
            "$ref": "#/parameters/limit"
          },
          {
            "$ref": "#/parameters/preferCount"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "items": {
                "$ref": "#/definitions/nodes"
              },
              "type": "array"
            }
          },
          "206": {
            "description": "Partial Content"
          }
        },
        "tags": [
          "nodes"
        ]
      },
      "post": {
        "parameters": [
          {
            "$ref": "#/parameters/body.nodes"
          },
          {
            "$ref": "#/parameters/select"
          },
          {
            "$ref": "#/parameters/preferPost"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "tags": [
          "nodes"
        ]
      },
      "delete": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.nodes.id"
          },
          {
            "$ref": "#/parameters/rowFilter.nodes.node_id"
          },
          {
            "$ref": "#/parameters/rowFilter.nodes.gateway_id"
          },
          {
            "$ref": "#/parameters/rowFilter.nodes.crop_name"
          },
          {
            "$ref": "#/parameters/rowFilter.nodes.created_at"
          },
          {
            "$ref": "#/parameters/preferReturn"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "nodes"
        ]
      },
      "patch": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.nodes.id"
          },
          {
            "$ref": "#/parameters/rowFilter.nodes.node_id"
          },
          {
            "$ref": "#/parameters/rowFilter.nodes.gateway_id"
          },
          {
            "$ref": "#/parameters/rowFilter.nodes.crop_name"
          },
          {
            "$ref": "#/parameters/rowFilter.nodes.created_at"
          },
          {
            "$ref": "#/parameters/body.nodes"
          },
          {
            "$ref": "#/parameters/preferReturn"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "nodes"
        ]
      }
    },
    "/gateways": {
      "get": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.gateways.id"
          },
          {
            "$ref": "#/parameters/rowFilter.gateways.gateway_id"
          },
          {
            "$ref": "#/parameters/rowFilter.gateways.gateway_key"
          },
          {
            "$ref": "#/parameters/rowFilter.gateways.owner_id"
          },
          {
            "$ref": "#/parameters/rowFilter.gateways.status"
          },
          {
            "$ref": "#/parameters/rowFilter.gateways.created_at"
          },
          {
            "$ref": "#/parameters/rowFilter.gateways.last_seen"
          },
          {
            "$ref": "#/parameters/select"
          },
          {
            "$ref": "#/parameters/order"
          },
          {
            "$ref": "#/parameters/range"
          },
          {
            "$ref": "#/parameters/rangeUnit"
          },
          {
            "$ref": "#/parameters/offset"
          },
          {
            "$ref": "#/parameters/limit"
          },
          {
            "$ref": "#/parameters/preferCount"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "items": {
                "$ref": "#/definitions/gateways"
              },
              "type": "array"
            }
          },
          "206": {
            "description": "Partial Content"
          }
        },
        "tags": [
          "gateways"
        ]
      },
      "post": {
        "parameters": [
          {
            "$ref": "#/parameters/body.gateways"
          },
          {
            "$ref": "#/parameters/select"
          },
          {
            "$ref": "#/parameters/preferPost"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "tags": [
          "gateways"
        ]
      },
      "delete": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.gateways.id"
          },
          {
            "$ref": "#/parameters/rowFilter.gateways.gateway_id"
          },
          {
            "$ref": "#/parameters/rowFilter.gateways.gateway_key"
          },
          {
            "$ref": "#/parameters/rowFilter.gateways.owner_id"
          },
          {
            "$ref": "#/parameters/rowFilter.gateways.status"
          },
          {
            "$ref": "#/parameters/rowFilter.gateways.created_at"
          },
          {
            "$ref": "#/parameters/rowFilter.gateways.last_seen"
          },
          {
            "$ref": "#/parameters/preferReturn"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "gateways"
        ]
      },
      "patch": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.gateways.id"
          },
          {
            "$ref": "#/parameters/rowFilter.gateways.gateway_id"
          },
          {
            "$ref": "#/parameters/rowFilter.gateways.gateway_key"
          },
          {
            "$ref": "#/parameters/rowFilter.gateways.owner_id"
          },
          {
            "$ref": "#/parameters/rowFilter.gateways.status"
          },
          {
            "$ref": "#/parameters/rowFilter.gateways.created_at"
          },
          {
            "$ref": "#/parameters/rowFilter.gateways.last_seen"
          },
          {
            "$ref": "#/parameters/body.gateways"
          },
          {
            "$ref": "#/parameters/preferReturn"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "gateways"
        ]
      }
    },
...
  },
  "definitions": {
...
    "nodes": {
      "required": [
        "id",
        "node_id"
      ],
      "properties": {
        "id": {
          "default": "gen_random_uuid()",
          "description": "Note:\nThis is a Primary Key.<pk/>",
          "format": "uuid",
          "type": "string"
        },
        "node_id": {
          "format": "text",
          "type": "string"
        },
        "gateway_id": {
          "description": "Note:\nThis is a Foreign Key to `gateways.id`.<fk table='gateways' column='id'/>",
          "format": "uuid",
          "type": "string"
        },
        "crop_name": {
          "format": "text",
          "type": "string"
        },
        "created_at": {
          "default": "now()",
          "format": "timestamp without time zone",
          "type": "string"
        }
      },
      "type": "object"
    },
    "gateways": {
      "required": [
        "id",
        "gateway_id",
        "gateway_key"
      ],
      "properties": {
        "id": {
          "default": "gen_random_uuid()",
          "description": "Note:\nThis is a Primary Key.<pk/>",
          "format": "uuid",
          "type": "string"
        },
        "gateway_id": {
          "format": "text",
          "type": "string"
        },
        "gateway_key": {
          "format": "text",
          "type": "string"
        },
        "owner_id": {
          "description": "Note:\nThis is a Foreign Key to `users.id`.<fk table='users' column='id'/>",
          "format": "uuid",
          "type": "string"
        },
        "status": {
          "default": "offline",
          "format": "text",
          "type": "string"
        },
        "created_at": {
          "default": "now()",
          "format": "timestamp without time zone",
          "type": "string"
        },
        "last_seen": {
          "format": "timestamp with time zone",
          "type": "string"
        }
      },
      "type": "object"
    },
...
  },
  "parameters": {
...
    "preferReturn": {
      "name": "Prefer",
      "description": "Preference",
      "required": false,
      "enum": [
        "return=representation",
        "return=minimal",
        "return=none"
      ],
      "in": "header",
      "type": "string"
    },
    "preferCount": {
      "name": "Prefer",
      "description": "Preference",
      "required": false,
      "enum": [
        "count=none"
      ],
      "in": "header",
      "type": "string"
    },
    "preferPost": {
      "name": "Prefer",
      "description": "Preference",
      "required": false,
      "enum": [
        "return=representation",
        "return=minimal",
        "return=none",
        "resolution=ignore-duplicates",
        "resolution=merge-duplicates"
      ],
      "in": "header",
      "type": "string"
    },
    "select": {
      "name": "select",
      "description": "Filtering Columns",
      "required": false,
      "in": "query",
      "type": "string"
    },
...
    "order": {
      "name": "order",
      "description": "Ordering",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "range": {
      "name": "Range",
      "description": "Limiting and Pagination",
      "required": false,
      "in": "header",
      "type": "string"
    },
    "rangeUnit": {
      "name": "Range-Unit",
      "description": "Limiting and Pagination",
      "required": false,
      "default": "items",
      "in": "header",
      "type": "string"
    },
    "offset": {
      "name": "offset",
      "description": "Limiting and Pagination",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "limit": {
      "name": "limit",
      "description": "Limiting and Pagination",
      "required": false,
      "in": "query",
      "type": "string"
    },
...
    "body.nodes": {
      "name": "nodes",
      "description": "nodes",
      "required": false,
      "in": "body",
      "schema": {
        "$ref": "#/definitions/nodes"
      }
    },
    "rowFilter.nodes.id": {
      "name": "id",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.nodes.node_id": {
      "name": "node_id",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.nodes.gateway_id": {
      "name": "gateway_id",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.nodes.crop_name": {
      "name": "crop_name",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.nodes.created_at": {
      "name": "created_at",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "body.gateways": {
      "name": "gateways",
      "description": "gateways",
      "required": false,
      "in": "body",
      "schema": {
        "$ref": "#/definitions/gateways"
      }
    },
    "rowFilter.gateways.id": {
      "name": "id",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.gateways.gateway_id": {
      "name": "gateway_id",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.gateways.gateway_key": {
      "name": "gateway_key",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.gateways.owner_id": {
      "name": "owner_id",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.gateways.status": {
      "name": "status",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.gateways.created_at": {
      "name": "created_at",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.gateways.last_seen": {
      "name": "last_seen",
      "required": false,
      "in": "query",
      "type": "string"
    },
...
  },
...
}